backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
clerk_backend_api 
svix
mysqlclient
BeautifulSoup4
playwright
langchain
langchain-google-genai
aiomysql
redis 
fastapi-cache2